Initial restructure for easier management/maintenance.
