+++ ghorgs/exceptions.py
@@ -1,0 +1,6 @@
+class GithubGraphQLError(Exception):
+    pass
+
+
+class UnexpectedResponseError(Exception):
+    pass
